rename js document
